+++ apx_core/src/url/canonical.rs
@@ -1,4 +1,4 @@
-//! Canonical URL
+//! Canonical URI
 
 use std::fmt;
 
@@ -10,8 +10,8 @@ use serde::{
 use thiserror::Error;
 
 use crate::{
-    ap_url::{is_ap_url, ApUrl},
-    http_url::HttpUrl,
+    ap_url::{is_ap_url as is_ap_uri, ApUrl as ApUri},
+    http_url::{HttpUrl as HttpUri},
     url::common::Origin,
 };
 
@@ -19,116 +19,116 @@ pub const GATEWAY_PATH_PREFIX: &str = "/.well-known/apgateway/";
 
 #[derive(Debug, Error)]
 #[error("{0}")]
-pub struct CanonicalUrlError(pub &'static str);
+pub struct CanonicalUriError(pub &'static str);
 
 #[derive(Clone, PartialEq)]
-pub enum CanonicalUrl {
-    Http(HttpUrl),
-    Ap(ApUrl),
+pub enum CanonicalUri {
+    Http(HttpUri),
+    Ap(ApUri),
 }
 
-pub fn with_gateway(ap_url: &ApUrl, gateway_url: &str) -> String {
-    format!("{}{}{}", gateway_url, GATEWAY_PATH_PREFIX, ap_url.to_did_url())
+pub fn with_gateway(ap_uri: &ApUri, gateway_base: &str) -> String {
+    format!("{}{}{}", gateway_base, GATEWAY_PATH_PREFIX, ap_uri.to_did_url())
 }
 
-impl CanonicalUrl {
-    pub fn parse(value: &str) -> Result<Self, CanonicalUrlError> {
-        let (url, _) = parse_url(value)?;
-        Ok(url)
+impl CanonicalUri {
+    pub fn parse(value: &str) -> Result<Self, CanonicalUriError> {
+        let (canonical_uri, _) = parse_url(value)?;
+        Ok(canonical_uri)
     }
 
-    pub fn parse_canonical(value: &str) -> Result<Self, CanonicalUrlError> {
-        let (url, maybe_gateway) = parse_url(value)?;
+    pub fn parse_canonical(value: &str) -> Result<Self, CanonicalUriError> {
+        let (canonical_uri, maybe_gateway) = parse_url(value)?;
         if maybe_gateway.is_some() {
-            return Err(CanonicalUrlError("URL is not canonical"));
+            return Err(CanonicalUriError("URI is not canonical"));
         };
-        Ok(url)
+        Ok(canonical_uri)
     }
 
     pub fn authority(&self) -> String {
         match self {
-            Self::Http(http_url) => http_url.authority().to_string(),
-            Self::Ap(ap_url) => ap_url.authority().to_string(),
+            Self::Http(http_uri) => http_uri.authority().to_string(),
+            Self::Ap(ap_uri) => ap_uri.authority().to_string(),
         }
     }
 
     pub fn origin(&self) -> Origin {
         match self {
-            Self::Http(http_url) => http_url.origin(),
-            Self::Ap(ap_url) => ap_url.origin(),
+            Self::Http(http_uri) => http_uri.origin(),
+            Self::Ap(ap_uri) => ap_uri.origin(),
         }
     }
 
-    pub fn to_http_url(&self, maybe_gateway: Option<&str>) -> Option<String> {
-        let url = match self {
-            Self::Http(http_url) => http_url.to_string(),
-            Self::Ap(ap_url) => {
+    pub fn to_http_uri(&self, maybe_gateway: Option<&str>) -> Option<String> {
+        let http_uri = match self {
+            Self::Http(http_uri) => http_uri.to_string(),
+            Self::Ap(ap_uri) => {
                 if let Some(gateway) = maybe_gateway {
-                    with_gateway(ap_url, gateway)
+                    with_gateway(ap_uri, gateway)
                 } else {
                     // Not enough context
                     return None;
                 }
             },
         };
-        Some(url)
+        Some(http_uri)
     }
 }
 
-impl fmt::Display for CanonicalUrl {
+impl fmt::Display for CanonicalUri {
     fn fmt(&self, formatter: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            Self::Http(http_url) => write!(formatter, "{}", http_url),
-            Self::Ap(ap_url) => write!(formatter, "{}", ap_url),
+            Self::Http(http_uri) => write!(formatter, "{}", http_uri),
+            Self::Ap(ap_uri) => write!(formatter, "{}", ap_uri),
         }
     }
 }
 
-fn get_canonical_ap_url(
-    http_url: HttpUrl,
-) -> Result<(ApUrl, String), CanonicalUrlError> {
-    let relative_http_url = http_url.to_relative();
-    let did_url = relative_http_url
+fn get_canonical_ap_uri(
+    http_uri: HttpUri,
+) -> Result<(ApUri, String), CanonicalUriError> {
+    let relative_http_uri = http_uri.to_relative();
+    let did_url = relative_http_uri
         .strip_prefix(GATEWAY_PATH_PREFIX)
-        .ok_or(CanonicalUrlError("invalid gateway URL"))?;
-    let ap_url = ApUrl::from_did_url(did_url)
-        .map_err(CanonicalUrlError)?;
-    let gateway = http_url.base();
-    Ok((ap_url, gateway))
+        .ok_or(CanonicalUriError("invalid gateway URI"))?;
+    let ap_uri = ApUri::from_did_url(did_url)
+        .map_err(CanonicalUriError)?;
+    let gateway = http_uri.base();
+    Ok((ap_uri, gateway))
 }
 
 pub fn parse_url(
     value: &str,
-) -> Result<(CanonicalUrl, Option<String>), CanonicalUrlError> {
+) -> Result<(CanonicalUri, Option<String>), CanonicalUriError> {
     let mut maybe_gateway = None;
-    let url = if is_ap_url(value) {
-        let ap_url = ApUrl::parse(value).map_err(CanonicalUrlError)?;
-        CanonicalUrl::Ap(ap_url)
+    let canonical_uri = if is_ap_uri(value) {
+        let ap_uri = ApUri::parse(value).map_err(CanonicalUriError)?;
+        CanonicalUri::Ap(ap_uri)
     } else {
-        let http_url = HttpUrl::parse(value).map_err(CanonicalUrlError)?;
-        if http_url.path().starts_with(GATEWAY_PATH_PREFIX) {
-            let (ap_url, gateway) = get_canonical_ap_url(http_url)?;
+        let http_uri = HttpUri::parse(value).map_err(CanonicalUriError)?;
+        if http_uri.path().starts_with(GATEWAY_PATH_PREFIX) {
+            let (ap_uri, gateway) = get_canonical_ap_uri(http_uri)?;
             maybe_gateway = Some(gateway);
-            CanonicalUrl::Ap(ap_url)
+            CanonicalUri::Ap(ap_uri)
         } else {
-            CanonicalUrl::Http(http_url)
+            CanonicalUri::Http(http_uri)
         }
     };
-    Ok((url, maybe_gateway))
+    Ok((canonical_uri, maybe_gateway))
 }
 
-impl<'de> Deserialize<'de> for CanonicalUrl {
+impl<'de> Deserialize<'de> for CanonicalUri {
     fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
         where D: Deserializer<'de>
     {
-        let url_str: String = Deserialize::deserialize(deserializer)?;
-        Self::parse(&url_str).map_err(DeserializerError::custom)
+        let value: String = Deserialize::deserialize(deserializer)?;
+        Self::parse(&value).map_err(DeserializerError::custom)
     }
 }
 
-pub fn is_same_origin(id_1: &str, id_2: &str) -> Result<bool, CanonicalUrlError> {
-    let id_1 = CanonicalUrl::parse(id_1)?;
-    let id_2 = CanonicalUrl::parse(id_2)?;
+pub fn is_same_origin(id_1: &str, id_2: &str) -> Result<bool, CanonicalUriError> {
+    let id_1 = CanonicalUri::parse(id_1)?;
+    let id_2 = CanonicalUri::parse(id_2)?;
     let is_same = id_1.origin() == id_2.origin();
     Ok(is_same)
 }
@@ -138,93 +138,93 @@ mod tests {
     use super::*;
 
     #[test]
-    fn test_http_url_from_http_url() {
-        let url_str = "https://social.example/users/test";
-        let http_url = HttpUrl::parse(url_str).unwrap();
-        let url = CanonicalUrl::Http(http_url);
-        let output = url.to_http_url(None).unwrap();
-        assert_eq!(output, url_str);
+    fn test_http_uri_from_http_uri() {
+        let input = "https://social.example/users/test";
+        let http_uri = HttpUri::parse(input).unwrap();
+        let canonical_uri = CanonicalUri::Http(http_uri);
+        let output = canonical_uri.to_http_uri(None).unwrap();
+        assert_eq!(output, input);
     }
 
     #[test]
-    fn test_http_url_from_ap_url() {
-        let url_str = "ap://did:key:z6MkvUie7gDQugJmyDQQPhMCCBfKJo7aGvzQYF2BqvFvdwx6/actor";
-        let ap_url = ApUrl::parse(url_str).unwrap();
-        let url = CanonicalUrl::Ap(ap_url);
+    fn test_http_uri_from_ap_uri() {
+        let input = "ap://did:key:z6MkvUie7gDQugJmyDQQPhMCCBfKJo7aGvzQYF2BqvFvdwx6/actor";
+        let ap_uri = ApUri::parse(input).unwrap();
+        let canonical_uri = CanonicalUri::Ap(ap_uri);
         let gateway = "https://gateway.example";
-        let output = url.to_http_url(Some(gateway)).unwrap();
+        let output = canonical_uri.to_http_uri(Some(gateway)).unwrap();
         let expected_output = "https://gateway.example/.well-known/apgateway/did:key:z6MkvUie7gDQugJmyDQQPhMCCBfKJo7aGvzQYF2BqvFvdwx6/actor";
         assert_eq!(output, expected_output);
     }
 
     #[test]
-    fn test_http_url_from_ap_url_no_gateway() {
-        let url_str = "ap://did:key:z6MkvUie7gDQugJmyDQQPhMCCBfKJo7aGvzQYF2BqvFvdwx6/actor";
-        let ap_url = ApUrl::parse(url_str).unwrap();
-        let url = CanonicalUrl::Ap(ap_url);
-        let maybe_output = url.to_http_url(None);
+    fn test_http_uri_from_ap_uri_no_gateway() {
+        let input = "ap://did:key:z6MkvUie7gDQugJmyDQQPhMCCBfKJo7aGvzQYF2BqvFvdwx6/actor";
+        let ap_uri = ApUri::parse(input).unwrap();
+        let canonical_uri = CanonicalUri::Ap(ap_uri);
+        let maybe_output = canonical_uri.to_http_uri(None);
         assert!(maybe_output.is_none());
     }
 
     #[test]
     fn test_parse_url_https() {
-        let url_str = "https://social.example/users/test";
-        let (url, maybe_gateway) = parse_url(url_str).unwrap();
-        assert!(matches!(url, CanonicalUrl::Http(_)));
+        let url = "https://social.example/users/test";
+        let (canonical_uri, maybe_gateway) = parse_url(url).unwrap();
+        assert!(matches!(canonical_uri, CanonicalUri::Http(_)));
         assert_eq!(maybe_gateway, None);
-        assert_eq!(url.to_string(), url_str);
+        assert_eq!(canonical_uri.to_string(), url);
     }
 
     #[test]
     fn test_parse_url_https_with_fragment() {
-        let url_str = "https://www.w3.org/ns/activitystreams#Public";
-        let (url, maybe_gateway) = parse_url(url_str).unwrap();
-        assert!(matches!(url, CanonicalUrl::Http(_)));
+        let url = "https://www.w3.org/ns/activitystreams#Public";
+        let (canonical_uri, maybe_gateway) = parse_url(url).unwrap();
+        assert!(matches!(canonical_uri, CanonicalUri::Http(_)));
         assert_eq!(maybe_gateway, None);
-        assert_eq!(url.to_string(), url_str);
+        assert_eq!(canonical_uri.to_string(), url);
     }
 
     #[test]
     fn test_parse_url_i2p() {
-        let url_str = "http://social.example.i2p/users/test";
-        let (url, maybe_gateway) = parse_url(url_str).unwrap();
-        assert!(matches!(url, CanonicalUrl::Http(_)));
+        let url = "http://social.example.i2p/users/test";
+        let (canonical_uri, maybe_gateway) = parse_url(url).unwrap();
+        assert!(matches!(canonical_uri, CanonicalUri::Http(_)));
         assert_eq!(maybe_gateway, None);
-        assert_eq!(url.to_string(), url_str);
+        assert_eq!(canonical_uri.to_string(), url);
     }
 
     #[test]
     fn test_parse_url_localhost() {
-        let url_str = "http://127.0.0.1:8380/users/test";
-        let (url, maybe_gateway) = parse_url(url_str).unwrap();
-        assert!(matches!(url, CanonicalUrl::Http(_)));
+        let url = "http://127.0.0.1:8380/users/test";
+        let (canonical_uri, maybe_gateway) = parse_url(url).unwrap();
+        assert!(matches!(canonical_uri, CanonicalUri::Http(_)));
         assert_eq!(maybe_gateway, None);
-        assert_eq!(url.to_string(), url_str);
+        assert_eq!(canonical_uri.to_string(), url);
     }
 
     #[test]
     fn test_parse_url_ap() {
-        let url_str = "ap://did:key:z6MkvUie7gDQugJmyDQQPhMCCBfKJo7aGvzQYF2BqvFvdwx6/actor";
-        let (url, maybe_gateway) = parse_url(url_str).unwrap();
-        assert!(matches!(url, CanonicalUrl::Ap(_)));
+        let url = "ap://did:key:z6MkvUie7gDQugJmyDQQPhMCCBfKJo7aGvzQYF2BqvFvdwx6/actor";
+        let (canonical_uri, maybe_gateway) = parse_url(url).unwrap();
+        assert!(matches!(canonical_uri, CanonicalUri::Ap(_)));
         assert_eq!(maybe_gateway, None);
-        assert_eq!(url.to_string(), url_str);
+        assert_eq!(canonical_uri.to_string(), url);
     }
 
     #[test]
     fn test_parse_url_ap_with_gateway() {
-        let url_str = "https://social.example/.well-known/apgateway/did:key:z6MkvUie7gDQugJmyDQQPhMCCBfKJo7aGvzQYF2BqvFvdwx6/actor";
-        let (url, maybe_gateway) = parse_url(url_str).unwrap();
-        assert!(matches!(url, CanonicalUrl::Ap(_)));
+        let url = "https://social.example/.well-known/apgateway/did:key:z6MkvUie7gDQugJmyDQQPhMCCBfKJo7aGvzQYF2BqvFvdwx6/actor";
+        let (canonical_uri, maybe_gateway) = parse_url(url).unwrap();
+        assert!(matches!(canonical_uri, CanonicalUri::Ap(_)));
         assert_eq!(maybe_gateway.as_deref(), Some("https://social.example"));
-        let expected_canonical_url = "ap://did:key:z6MkvUie7gDQugJmyDQQPhMCCBfKJo7aGvzQYF2BqvFvdwx6/actor";
-        assert_eq!(url.to_string(), expected_canonical_url);
+        let expected_canonical_uri = "ap://did:key:z6MkvUie7gDQugJmyDQQPhMCCBfKJo7aGvzQYF2BqvFvdwx6/actor";
+        assert_eq!(canonical_uri.to_string(), expected_canonical_uri);
     }
 
     #[test]
     fn test_parse_url_ap_with_gateway_unsupported_did() {
-        let url_str = "https://social.example/.well-known/apgateway/did:example:123456";
-        let error = parse_url(url_str).err().unwrap();
+        let url = "https://social.example/.well-known/apgateway/did:example:123456";
+        let error = parse_url(url).err().unwrap();
         assert_eq!(error.to_string(), "invalid 'ap' URL");
     }