+++ apx_sdk/CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Renamed `CanonicalUrl` type to `CanonicalUri`.
+
 ## [0.18.0] - 2025-10-08
 
 ### Added